Simulation by using Threading and MPI
This simulation creditted to (http://scipy-cookbook.readthedocs.io/items/BrownianMotion.html) where the main brownian function is not actually a part of the library in Python.

Side note: My first mistake during doing this by misunderstanding the function as a part of Python Library as shown in the original work in above.

1-BMotionBasic:
 One particle only involved and moved randomly in two dimension
2-BMotion2Particle:
3 particles are randomly moving in the same dimesnion by using Threading method. Collision between the particles are not being considered in this case
3-BMotionMPI:
Using Message Passing Interface, to calculate the position of the particles